requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.32.0
google-generativeai==0.8.3
streamlit==1.39.0
python-dotenv==1.0.1
mcp==0.0.6
